requirements.txt
git+https://github.com/yondonfu/diffusers.git@kandinsky-ddim-inverse
Pillow
transformers
accelerate
huggingface_hub
fastapi